Gemfile.lock not recommended for Gems
I had a problem including Bgg in my Gemfile because it was trying to install a version of the git-gem that didn't know my system git version even existed.

``` [WARNING] The git gem requires git 1.6.0.0 or later, but only found 2.1.4. You should probably upgrade.```

Removing the Gemfile.lock file let bundler recalculate dependency tree and afterwards I could install it via a Gemfile.

If specific version of a gem must be tied in, that should be set in the gemspec dependencies.